# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API test-backend-developer",
        "description": "Listado de URI's de la API test-backend-developer",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/createDetalleVenta": {
            "post": {
                "tags": [
                    "DetalleVenta"
                ],
                "summary": "Crear un nuevo detalle de venta",
                "operationId": "App\\Http\\Controllers\\DetalleVentaController::createDetalleVenta",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_venta": {
                                        "description": "Crear un nuevo detalle de venta",
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "id_producto": {
                                        "description": "Crear un nuevo detalle de venta",
                                        "type": "number",
                                        "example": "2"
                                    },
                                    "cantidad": {
                                        "description": "Crear un nuevo detalle de venta",
                                        "type": "number",
                                        "example": "30"
                                    },
                                    "precio_unitario": {
                                        "description": "Crear un nuevo detalle de venta",
                                        "type": "number",
                                        "example": "567.00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Detalle de Venta creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Crear un nuevo detalle de venta",
                                            "type": "string",
                                            "example": "Detalle de Venta creado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Crear un nuevo detalle de venta",
                                            "type": "string",
                                            "example": "Error en los datos enviados"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}